+++ src/main/java/com/mylearning/Student/StudentRepository.java
@@ -2,8 +2,8 @@ package com.mylearning.Student;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepository extends JpaRepository<Student,Integer> {
+public interface StudentRepository extends JpaRepository<Student,Long> {
     boolean existsStudentByEmail(String email);
-    boolean existsStudentById(Integer studentId);
+    boolean existsStudentById(Long studentId);
 
 }
